+++ test/erlfmt_SUITE_data/ignore_format_old.erl
@@ -1,6 +1,6 @@
--module(ignore_format_many).
+-module(ignore_format_old).
 
-%%% erlfmt:ignore-begin
+%%% erlfmt-ignore
 -define(DELTA_MATRIX, [
     [0,   0,   0,   0,   0,   0],
     [0, -16,   0,   0,   0,   0],
@@ -9,20 +9,18 @@
     [0, -16,   0,   0, -14,   0],
     [0,   0,  15,   0,   0,   0]
 ]).
--define(ALSO_IGNORED, [  1,  2,  3]).
 
-%%% erlfmt:ignore-end
 -define(DELTA_MATRIX_FORMATTED, [
-    [0, 0, 0, 0, 0, 0],
-    [0, -16, 0, 0, 0, 0],
-    [0, 0, 15, 0, 0, 0],
-    [0, 0, 0, 6, 0, 0],
-    [0, -16, 0, 0, -14, 0],
-    [0, 0, 15, 0, 0, 0]
+    [0,   0,   0,   0,   0,   0],
+    [0, -16,   0,   0,   0,   0],
+    [0,   0,  15,   0,   0,   0],
+    [0,   0,   0,   6,   0,   0],
+    [0, -16,   0,   0, -14,   0],
+    [0,   0,  15,   0,   0,   0]
 ]).
 
 %% some comment
-%%erlfmt:ignore %
+%%erlfmt-ignore %
 %% another comment
 gen_part_decode_funcs({constructed,bif},TypeName,
               {_Name,parts,Tag,_Type}) ->
@@ -35,18 +33,8 @@ gen_part_decode_funcs({constructed,bif},TypeName,
           "      Res",nl,
           "  end"]).
 
-% erlfmt:ignore-begin I like the comments next to the statement
+% erlfmt-ignore I like the comment next to the statement
 f() -> ok. % this is ok
-g() -> ok. % this is also ok
-% erlfmt:ignore-end I'm done with this style
-
-% erlfmt:ignore-begin Here we go again...
-a1() -> ok. % again
-a2() -> ok. % ditto
-% erlfmt:ignore-end Now I'm really done
-
-% blah
-h() -> ok.
 
 %% TODO write emit
-emit(S) -> ok.
+emit(S) ->   ok.